refactor: navigation dropdown
